PHP8 support
I'm running this code with PHP8.2 and it fails with following error message:

Unexpected error, the message was : explode(): Argument #2 ($string) must be of type string, array given in ./ezpublish_legacy/lib/ezutils/classes/ezini.php on line 1751

As you can see in the original code, sometimes this function gets an array for the $path variable. In that case, the original code was picking the first element of the array if it exists. In case of an empty array, $path is still an array. 

The PHP function explode `explode( '/', $path );` expects a string for $path. PHP7 and lower just return FALSE. PHP8 triggers an error and stops the execution.

My code changes just make sure that $path is a string and it returns 'undefined' in case of an empty array for the $path variable.